requirements.txt
openai
spacy
pandas
pydantic
weasyprint
streamlit
rapidfuzz
openpyxl
python-docx
markdown-pdf
streamlit_quill
streamlit_option_menu
beautifulsoup4
python-docx
PyPDF2